Added appConfig scaffolding
-At first I had alot of problems. No matter what I did I could not get
the values set in the App.config. But I found out that the App.Config
was different from the debug and release cause there were way too many
versions of SmartRestaurant. So with tomorrow, we have to make sure that
the appConfig we set is the one that is going to be used. The solution
to the problem was to open showallfiles and look for anything with
.config. Then look for the one in debug.
-I made a test project and used it to check if it can get the correct
strings. We will continue to use this in the future.
-I removed the SquirelTemp from git. Hopefully Yewo did the same
otherwise it would be so cluterful
-New order just had a slight change of variable names