DO NOT PULL FROM HERE
-There was too much, way too much to finish in one sitting.
FirebaseDatacontext was actually hell. I thought it was bad in
QRCashless but it was nothing compared. But let me try breaking down
progress

-The Menu aggregate was introduced. This was also done in QR. We also
introduced an exception and other works that were found in QR or were
created as a consequence of the copy. Please not that the kernel will
have to remove most of these copies.
-I plan on having OfflineDataHelpers become the service in charge of all
things offline.
-Two xaml files were edited to have the MenuService included in them.

Now for the pain
-So in FirebaseDataContext. Alot of stuff that should have been in a
different thing from firebase was there, but at the same time was
heavily coupled with the things inside firebase. So it was difficult
untieing the mess. It still is. Not only was it coupled with
firebase functionality but also OfflineContext. So As it stands there
are alot that aren't connected and are still loose within both
firebase and dataService.
-we also have a problem where DataService required firebaseDataContext
directly. I am hoping I can make a DatabaseService that will encapsulate
these new methods and simply get its data from FirebaseServices as an
inherited property. But that's a problem for future Abel.