Investigation
Error was most likely caused by an online order which was incorrectly formatted and thus caused errors on deserialization.

I edited the App.xaml.cs to see if the error log was causing an exception which was causing the rest of the code not to be executed and the email of the error log not to be sent.