OfflineData and firebase
-A method to have data saved regardless of what data type it is so that
the TCP Server can be used.
-Firebase tests were edited. all of them were working, and now its like
there are situations where you can't get data that is a entity within an
aggregate. It gets read as a jproperty instead of an jobject and so it
throws weird errors cause its trying to parse something into something
but its not really possible with jproperty but rather with jobject. We
need to either fix that so it can do a better job telling what is what,
or have it fixed that it accepts only certain types. ie, it only takes
in lists of OrderItems and MenuItems